backend/requirements.txt
# ── Backend Dependencies ─────────────────────────────────────────
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Validation
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Async tasks
celery==5.4.0
redis==5.2.1

# CORS & security
python-multipart==0.0.20

# Auth
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# File upload
python-multipart
